fix: 랜덤값 만큼 이동하는게 아닌 1씩 이동하도록 변경, 테스트 케이스 변경
