+++ src/main/java/racing/domain/Car.java
@@ -1,38 +1,19 @@
 package racing.domain;
 
-import java.util.Random;
-
-import static racing.domain.CarRacingProperty.*;
+import static racing.domain.CarRacingProperty.MIN_MOVABLE_DISTANCE;
 
 public class Car {
 
-    private final Random random;
+    private int distance = 0;
 
-    private int totalDistance = 0;
-
-    public Car() {
-        random = new Random();
-    }
-
-    public int move() {
-        return move(null);
-    }
-
-    public int move(RaceFunction function) {
-        int distance = getRandomDistance();
-        distance = distance >= MIN_MOVABLE_DISTANCE ? distance : 0;
-        totalDistance += distance;
-        if (function != null) {
-            function.compute(distance, totalDistance);
+    public void move(int distance) {
+        if (distance >= MIN_MOVABLE_DISTANCE) {
+            this.distance++;
         }
+    }
+
+    public int getDistance() {
         return distance;
     }
 
-    public int getTotalDistance() {
-        return totalDistance;
-    }
-
-    public int getRandomDistance() {
-        return random.nextInt(MAX_RANDOM_DISTANCE - MIN_RANDOM_DISTANCE + 1) + MIN_RANDOM_DISTANCE;
-    }
 }
