+++ src/main/java/racing/domain/CarRacing.java
@@ -2,16 +2,22 @@ package racing.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
+
+import static racing.domain.CarRacingProperty.MAX_RANDOM_DISTANCE;
+import static racing.domain.CarRacingProperty.MIN_RANDOM_DISTANCE;
 
 public class CarRacing {
 
     private final int carCount;
     private final int tryCount;
 
+    private Random random;
     private List<Car> cars;
 
     private boolean complete = false;
     private int racingCount = 0;
+
 
     public CarRacing(int carCount, int tryCount) {
         this.carCount = carCount;
@@ -51,6 +57,7 @@ public class CarRacing {
             return;
         }
 
+        random = new Random();
         cars = new ArrayList<>(carCount);
         for (int i = 0; i < carCount; i++) {
             cars.add(new Car());
@@ -59,8 +66,14 @@ public class CarRacing {
 
     private void tryMoveCars(RaceFunction mappingFunction) {
         for (Car car : cars) {
-            car.move(mappingFunction);
+            car.move(getRandomDistance());
+            if (mappingFunction != null) {
+                mappingFunction.compute(car.getDistance());
+            }
         }
     }
 
+    public int getRandomDistance() {
+        return random.nextInt(MAX_RANDOM_DISTANCE - MIN_RANDOM_DISTANCE + 1) + MIN_RANDOM_DISTANCE;
+    }
 }
