+++ src/main/java/racing/domain/RaceFunction.java
@@ -2,5 +2,5 @@ package racing.domain;
 
 @FunctionalInterface
 public interface RaceFunction {
-    void compute(int movedDistance, int totalDistance);
+    void compute(int totalDistance);
 }
