+++ src/test/java/racing/domain/CarTest.java
@@ -11,28 +11,20 @@ class CarTest {
     @DisplayName("차 이동 시 최소거리 이상 확인")
     void move() {
         Car car = new Car();
-        while (car.move() > 0) {
-            assertThat(car.getTotalDistance()).isGreaterThan(CarRacingProperty.MIN_MOVABLE_DISTANCE);
-            break;
-        }
+        int distance = car.getDistance();
+        car.move(3);
+        assertThat(car.getDistance()).isEqualTo(distance);
+        car.move(4);
+        assertThat(car.getDistance()).isEqualTo(distance + 1);
     }
 
     @Test
     @DisplayName("최종 이동거리 확인")
     void getTotalDistance() {
         Car car = new Car();
-        int distance1 = car.move();
-        int distance2 = car.move();
-        assertThat(car.getTotalDistance()).isEqualTo(distance1 + distance2);
+        car.move(9);
+        car.move(9);
+        assertThat(car.getDistance()).isEqualTo(2);
     }
 
-    @Test
-    @DisplayName("이동거리 랜덤 범위 확인")
-    void getRandomDistance() {
-        for (int i = 0; i < 10; i++) {
-            Car car = new Car();
-            int distance = car.getRandomDistance();
-            assertThat(distance).isBetween(CarRacingProperty.MIN_RANDOM_DISTANCE, CarRacingProperty.MAX_RANDOM_DISTANCE);
-        }
-    }
 }
